reorganize AppServices
