+++ src/AppServices/ApiManagementPage.test.tsx
@@ -1,12 +1,12 @@
-import { render, waitForI18n } from "../../test-utils";
 import { composeStories } from "@storybook/testing-react";
+import { render, waitForI18n } from "../test-utils";
 import * as stories from "./APIManagementPage.stories";
 
-const { Example } = composeStories(stories);
+const { APIManagementPage } = composeStories(stories);
 
 describe("ApiManagementPage", () => {
   it("renders", async () => {
-    const comp = render(<Example />);
+    const comp = render(<APIManagementPage />);
     await waitForI18n(comp);
     expect(comp.baseElement).toMatchSnapshot();
   });
