--- src/AppServices/DataSciencePage/index.ts
@@ -1,1 +1,0 @@
-export * from "./DataSciencePage";
