--- src/AppServices/KafkaPage/component/index.ts
@@ -1,1 +1,0 @@
-export * from "./KafkaInstanceCapacityTable";
