--- src/AppServices/KafkaPage/index.ts
@@ -1,2 +1,0 @@
-export * from "./KafkaPage";
-export * from "./KafkaPageV2";
