+++ src/AppServices/KafkaPageV2.test.tsx
@@ -1,12 +1,12 @@
-import { render, waitForI18n } from "../../test-utils";
 import { composeStories } from "@storybook/testing-react";
-import * as stories from "./KafkaPage.stories";
+import { render, waitForI18n } from "../test-utils";
+import * as stories from "./KafkaPageV2.stories";
 
-const { Example } = composeStories(stories);
+const { KafkaPageV2 } = composeStories(stories);
 
 describe("KafkaPage", () => {
   it("renders", async () => {
-    const comp = render(<Example />);
+    const comp = render(<KafkaPageV2 />);
     await waitForI18n(comp);
     expect(comp.baseElement).toMatchSnapshot();
   });
