+++ src/AppServices/OverviewPage.tsx
@@ -9,7 +9,6 @@ import {
   CardTitle,
   Grid,
   Label,
-  LabelGroup,
   Stack,
   StackItem,
   Title,
@@ -20,38 +19,35 @@ import { useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
 import {
   AppServicesOverviewIconPattern,
-  LogoRedHatOpenShiftDatabaseAccessAStandardRgb,
   LogoRedHatOpenShiftApiManagementAStandardRgb,
-  LogoRedHatOpenShiftConnectorsBStandardRgb,
+  LogoRedHatOpenShiftDatabaseAccessAStandardRgb,
   LogoRedHatOpenShiftDataScienceAStandardRgb,
   LogoRedHatOpenShiftServiceRegistryAStandardRgb,
   LogoRedHatOpenShiftStreamsForApacheKafkaAStandardRgbPng,
-} from "../../images";
+} from "../images";
 import {
   MarketingPageBanner,
   MarketingPageHero,
   MarketingPageSection,
-} from "../components";
+} from "./components";
 
-export type OverviewPageV2Props = {
+export type OverviewPageProps = {
   toKafkaHref: string;
   toServiceRegistryHref: string;
-  toConnectorsHref: string;
 };
 
-export const OverviewPageV2: FunctionComponent<OverviewPageV2Props> = ({
+export const OverviewPage: FunctionComponent<OverviewPageProps> = ({
   toKafkaHref,
   toServiceRegistryHref,
-  toConnectorsHref,
 }) => {
   const { t } = useTranslation();
   return (
     <>
       <MarketingPageHero
-        title={t("overview-v2:heroTitle")}
-        tagLine={t("overview-v2:heroTagline")}
-        description={t("overview-v2:heroDescription")}
-        description2={t("overview-v2:heroDescription2")}
+        title={t("overview:heroTitle")}
+        tagLine={t("overview:heroTagline")}
+        description={t("overview:heroDescription")}
+        description2={t("overview:heroDescription2")}
         heroImage={AppServicesOverviewIconPattern}
         heroImageSize={678}
         cta={
@@ -67,12 +63,12 @@ export const OverviewPageV2: FunctionComponent<OverviewPageV2Props> = ({
               />
             )}
           >
-            {t("overview-v2:heroCallToActionButton")}
+            {t("overview:heroCallToActionButton")}
           </Button>
         }
         variant="dark"
       />
-      <MarketingPageBanner>{t("overview-v2:banner")}</MarketingPageBanner>
+      <MarketingPageBanner>{t("overview:banner")}</MarketingPageBanner>
       <MarketingPageSection>
         <Grid md={6} lg={3} hasGutter>
           {/* API Management card */}
@@ -88,21 +84,14 @@ export const OverviewPageV2: FunctionComponent<OverviewPageV2Props> = ({
             </CardHeader>
             <CardTitle>
               <Title headingLevel="h2" size="xl">
-                {t("overview-v2:rhoamTitle")}
+                {t("overview:rhoamTitle")}
               </Title>
             </CardTitle>
             <CardBody>
               <Stack hasGutter>
-                <StackItem className="pf-u-mb-lg">
-                  <LabelGroup>
-                    <Label color="blue">
-                      {t("overview-v2:applicationService")}
-                    </Label>
-                  </LabelGroup>
-                </StackItem>
-                <StackItem>{t("overview-v2:rhoamMainText")}</StackItem>
+                <StackItem>{t("overview:rhoamMainText")}</StackItem>
                 <StackItem className="pf-u-color-200">
-                  {t("overview-v2:rhoamSecondaryText")}{" "}
+                  {t("overview:rhoamSecondaryText")}{" "}
                   <Button
                     data-testid="cardRHOAM-linkOpenShift"
                     isInline
@@ -126,212 +115,7 @@ export const OverviewPageV2: FunctionComponent<OverviewPageV2Props> = ({
                 target="_blank"
                 href="https://developers.redhat.com/products/rhoam/getting-started"
               >
-                {t("overview-v2:getStarted")}{" "}
-                <ExternalLinkAltIcon className="pf-u-ml-sm" />
-              </Button>
-            </CardFooter>
-          </Card>
-
-          {/* Kafka card */}
-          <Card ouiaId="card-rhosak">
-            <CardHeader>
-              <CardHeaderMain>
-                <img
-                  src={LogoRedHatOpenShiftStreamsForApacheKafkaAStandardRgbPng}
-                  alt="Red Hat OpenShift Streams for Apache Kafka logo"
-                  style={{ height: "50px" }}
-                />
-              </CardHeaderMain>
-            </CardHeader>
-            <CardTitle>
-              <Title headingLevel="h2" size="xl">
-                {t("overview-v2:rhosakTitle")}
-              </Title>
-            </CardTitle>
-            <CardBody>
-              <Stack hasGutter>
-                <StackItem className="pf-u-mb-lg">
-                  <LabelGroup>
-                    <Label color="blue">
-                      {t("overview-v2:applicationService")}
-                    </Label>
-                    <Label color="green">{t("overview-v2:dataService")}</Label>
-                  </LabelGroup>
-                </StackItem>
-                <StackItem>{t("overview-v2:rhosakMainText")}</StackItem>
-                <StackItem className="pf-u-color-200">
-                  {t("overview-v2:rhosakSecondaryText")}
-                </StackItem>
-              </Stack>
-            </CardBody>
-            <CardFooter>
-              <Button
-                data-testid="cardRHOSAK-buttonCreateKafka"
-                variant={ButtonVariant.secondary}
-                component={(props) => <Link {...props} to={toKafkaHref} />}
-                ouiaId="button-create"
-              >
-                {t("overview-v2:rhosakCallToActionButton")}
-              </Button>
-            </CardFooter>
-          </Card>
-
-          {/* Service Registry card */}
-          <Card>
-            <CardHeader>
-              <CardHeaderMain>
-                <img
-                  src={LogoRedHatOpenShiftServiceRegistryAStandardRgb}
-                  alt="Red Hat OpenShift Service Registry logo"
-                  style={{ height: "50px" }}
-                />
-              </CardHeaderMain>
-            </CardHeader>
-            <CardTitle>
-              <Title headingLevel="h2" size="xl">
-                {t("overview-v2:rhosrTitle")}
-              </Title>
-            </CardTitle>
-            <CardBody>
-              <Stack hasGutter>
-                <StackItem className="pf-u-mb-lg">
-                  <LabelGroup>
-                    <Label color="blue">
-                      {t("overview-v2:applicationService")}
-                    </Label>
-                  </LabelGroup>
-                </StackItem>
-                <StackItem>{t("overview-v2:rhosrMainText")}</StackItem>
-                <StackItem className="pf-u-color-200">
-                  {t("overview-v2:rhosrSecondaryText")}
-                </StackItem>
-              </Stack>
-            </CardBody>
-            <CardFooter>
-              <Button
-                data-testid="cardRHOSR-buttonCTA"
-                variant={ButtonVariant.secondary}
-                component={(props) => (
-                  <Link
-                    {...props}
-                    data-testid="hero-buttonTryKafka"
-                    to={toServiceRegistryHref}
-                  />
-                )}
-              >
-                {t("overview-v2:rhosrCallToActionButton")}
-              </Button>
-            </CardFooter>
-          </Card>
-
-          {/* Connectors card */}
-          <Card>
-            <CardHeader>
-              <CardHeaderMain>
-                <img
-                  src={LogoRedHatOpenShiftConnectorsBStandardRgb}
-                  alt="Red Hat OpenShift Connectors logo"
-                  style={{ height: "50px" }}
-                />
-              </CardHeaderMain>
-            </CardHeader>
-            <CardTitle>
-              <Title headingLevel="h2" size="xl">
-                {t("overview-v2:rhocTitle")}
-              </Title>
-            </CardTitle>
-            <CardBody>
-              <Stack hasGutter>
-                <StackItem className="pf-u-mb-lg">
-                  <LabelGroup>
-                    <Label color="blue">
-                      {t("overview-v2:applicationService")}
-                    </Label>
-                    <Label>{t("overview-v2:servicePreview")}</Label>
-                  </LabelGroup>
-                </StackItem>
-                <StackItem>{t("overview-v2:rhocMainText")}</StackItem>
-                <StackItem className="pf-u-color-200">
-                  {t("overview-v2:rhocSecondaryText")}{" "}
-                  <Button
-                    data-testid="cardRHOC-linkOpenShift"
-                    isInline
-                    variant={ButtonVariant.link}
-                    component="a"
-                    target="_blank"
-                    href="https://openshift.com"
-                  >
-                    OpenShift
-                    <ExternalLinkAltIcon className="pf-u-ml-xs" />
-                  </Button>
-                  {t("overview-v2:rhocSecondaryTextContinued")}
-                </StackItem>
-              </Stack>
-            </CardBody>
-            <CardFooter>
-              <Button
-                data-testid="cardRHOC-buttonCTA"
-                variant={ButtonVariant.secondary}
-                target="_blank"
-                component={(props) => <Link {...props} to={toConnectorsHref} />}
-              >
-                {t("overview-v2:getStarted")}
-              </Button>
-            </CardFooter>
-          </Card>
-
-          {/* Database access card */}
-          <Card>
-            <CardHeader>
-              <CardHeaderMain>
-                <img
-                  src={LogoRedHatOpenShiftDatabaseAccessAStandardRgb}
-                  alt="Red Hat OpenShift Data Science logo"
-                  style={{ height: "50px" }}
-                />
-              </CardHeaderMain>
-            </CardHeader>
-            <CardTitle>
-              <Title headingLevel="h2" size="xl">
-                {t("overview-v2:dbaasTitle")}
-              </Title>
-            </CardTitle>
-            <CardBody>
-              <Stack hasGutter>
-                <StackItem className="pf-u-mb-lg">
-                  <LabelGroup>
-                    <Label color="green">{t("overview:dataService")}</Label>
-                    <Label>{t("overview-v2:servicePreview")}</Label>
-                  </LabelGroup>
-                </StackItem>
-                <StackItem>{t("overview-v2:dbaasMainText")}</StackItem>
-                <StackItem className="pf-u-color-200">
-                  {t("overview-v2:dbaasSecondaryText")}{" "}
-                  <Button
-                    data-testid="cardRHODS-linkOpenShift"
-                    isInline
-                    variant={ButtonVariant.link}
-                    component="a"
-                    target="_blank"
-                    href="https://openshift.com"
-                  >
-                    {" "}
-                    OpenShift
-                    <ExternalLinkAltIcon className="pf-u-ml-xs" />
-                  </Button>
-                  .
-                </StackItem>
-              </Stack>
-            </CardBody>
-            <CardFooter>
-              <Button
-                data-testid="cardRHODS-buttonCTA"
-                variant={ButtonVariant.secondary}
-                component="a"
-                target="_blank"
-                href="https://console.redhat.com/application-services/databases"
-              >
-                {t("overview-v2:learnMore")}{" "}
+                {t("overview:getStarted")}{" "}
                 <ExternalLinkAltIcon className="pf-u-ml-sm" />
               </Button>
             </CardFooter>
@@ -350,20 +134,17 @@ export const OverviewPageV2: FunctionComponent<OverviewPageV2Props> = ({
             </CardHeader>
             <CardTitle>
               <Title headingLevel="h2" size="xl">
-                {t("overview-v2:rhodsTitle")}
+                {t("overview:rhodsTitle")}
               </Title>
             </CardTitle>
             <CardBody>
               <Stack hasGutter>
                 <StackItem className="pf-u-mb-lg">
-                  <LabelGroup>
-                    <Label color="green">{t("overview-v2:dataService")}</Label>
-                    <Label>{t("overview-v2:fieldTrial")}</Label>
-                  </LabelGroup>
+                  <Label>{t("overview:fieldTrial")}</Label>
                 </StackItem>
-                <StackItem>{t("overview-v2:rhodsMainText")}</StackItem>
+                <StackItem>{t("overview:rhodsMainText")}</StackItem>
                 <StackItem className="pf-u-color-200">
-                  {t("overview-v2:rhodsSecondaryText")}{" "}
+                  {t("overview:rhodsSecondaryText")}{" "}
                   <Button
                     data-testid="cardRHODS-linkOpenShift"
                     isInline
@@ -388,7 +169,138 @@ export const OverviewPageV2: FunctionComponent<OverviewPageV2Props> = ({
                 component="a"
                 target="_blank"
               >
-                {t("overview-v2:tryIt")}
+                {t("overview:tryIt")}
+              </Button>
+            </CardFooter>
+          </Card>
+
+          {/* Service Registry card */}
+          <Card>
+            <CardHeader>
+              <CardHeaderMain>
+                <img
+                  src={LogoRedHatOpenShiftServiceRegistryAStandardRgb}
+                  alt="Red Hat OpenShift Service Registry logo"
+                  style={{ height: "50px" }}
+                />
+              </CardHeaderMain>
+            </CardHeader>
+            <CardTitle>
+              <Title headingLevel="h2" size="xl">
+                {t("overview:rhosrTitle")}
+              </Title>
+            </CardTitle>
+            <CardBody>
+              <Stack hasGutter>
+                <StackItem>{t("overview:rhosrMainText")}</StackItem>
+                <StackItem className="pf-u-color-200">
+                  {t("overview:rhosrSecondaryText")}
+                </StackItem>
+              </Stack>
+            </CardBody>
+            <CardFooter>
+              <Button
+                data-testid="cardRHOSR-buttonCTA"
+                variant={ButtonVariant.secondary}
+                component={(props) => (
+                  <Link
+                    {...props}
+                    data-testid="hero-buttonTryKafka"
+                    to={toServiceRegistryHref}
+                  />
+                )}
+              >
+                {t("overview:rhosrCallToActionButton")}
+              </Button>
+            </CardFooter>
+          </Card>
+
+          {/* Kafka card */}
+          <Card ouiaId="card-rhosak">
+            <CardHeader>
+              <CardHeaderMain>
+                <img
+                  src={LogoRedHatOpenShiftStreamsForApacheKafkaAStandardRgbPng}
+                  alt="Red Hat OpenShift Streams for Apache Kafka logo"
+                  style={{ height: "50px" }}
+                />
+              </CardHeaderMain>
+            </CardHeader>
+            <CardTitle>
+              <Title headingLevel="h2" size="xl">
+                {t("overview:rhosakTitle")}
+              </Title>
+            </CardTitle>
+            <CardBody>
+              <Stack hasGutter>
+                <StackItem>{t("overview:rhosakMainText")}</StackItem>
+                <StackItem className="pf-u-color-200">
+                  {t("overview:rhosakSecondaryText")}
+                </StackItem>
+              </Stack>
+            </CardBody>
+            <CardFooter>
+              <Button
+                data-testid="cardRHOSAK-buttonCreateKafka"
+                variant={ButtonVariant.secondary}
+                component={(props) => <Link {...props} to={toKafkaHref} />}
+                ouiaId="button-create"
+              >
+                {t("overview:rhosakCallToActionButton")}
+              </Button>
+            </CardFooter>
+          </Card>
+
+          {/* Database access card */}
+          <Card>
+            <CardHeader>
+              <CardHeaderMain>
+                <img
+                  src={LogoRedHatOpenShiftDatabaseAccessAStandardRgb}
+                  alt="Red Hat OpenShift Data Science logo"
+                  style={{ height: "50px" }}
+                />
+              </CardHeaderMain>
+            </CardHeader>
+            <CardTitle>
+              <Title headingLevel="h2" size="xl">
+                {t("overview:dbaasTitle")}
+              </Title>
+            </CardTitle>
+            <CardBody>
+              <Stack hasGutter>
+                <StackItem className="pf-u-mb-lg">
+                  <Label>{t("overview:servicePreview")}</Label>
+                </StackItem>
+                <StackItem>{t("overview:dbaasMainText")}</StackItem>
+                <StackItem className="pf-u-color-200">
+                  {t("overview:dbaasSecondaryText")}{" "}
+                  <Button
+                    data-testid="cardRHODS-linkOpenShift"
+                    isInline
+                    variant={ButtonVariant.link}
+                    component="a"
+                    target="_blank"
+                    href="https://openshift.com"
+                  >
+                    {" "}
+                    OpenShift
+                    <ExternalLinkAltIcon className="pf-u-ml-xs" />
+                  </Button>
+                  .
+                </StackItem>
+              </Stack>
+            </CardBody>
+            <CardFooter>
+              <Button
+                data-testid="cardRHODS-buttonCTA"
+                variant={ButtonVariant.secondary}
+                component="a"
+                target="_blank"
+                href="https://console.redhat.com/application-services/databases"
+              >
+                {t("overview:learnMore")}{" "}
+                <ExternalLinkAltIcon className="pf-u-ml-sm" />
               </Button>
             </CardFooter>
           </Card>
