--- src/AppServices/OverviewPage/OverviewPage.stories.tsx
@@ -1,16 +1,0 @@
-import { ComponentMeta, ComponentStory } from "@storybook/react";
-import { OverviewPage } from "./OverviewPage";
-
-export default {
-  component: OverviewPage,
-  args: {
-    toKafkaHref: "/to-kafka",
-    toServiceRegistryHref: "/to-service",
-  },
-} as ComponentMeta<typeof OverviewPage>;
-
-const Template: ComponentStory<typeof OverviewPage> = (args) => (
-  <OverviewPage {...args} />
-);
-
-export const Example = Template.bind({});
