--- src/AppServices/OverviewPage/index.ts
@@ -1,2 +1,0 @@
-export * from "./OverviewPage";
-export * from "./OverviewPageV2";
