--- src/AppServices/components/MarketingPage/index.ts
@@ -1,1 +1,0 @@
-export * from "./MarketingPage";
