+++ src/AppServices/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./MarketingPage";
+export * from "./KafkaInstanceCapacityTable";
