+++ src/AppServices/index.ts
@@ -1,4 +1,4 @@
-export * from "./APIManagement";
+export * from "./APIManagementPage";
 export * from "./DataSciencePage";
 export * from "./OverviewPage";
 export * from "./KafkaPage";
